+++ CHANGES.md
@@ -4,6 +4,7 @@
 * updated to use Bitbucket API 2.0 where appropriate
 * refactor API calls with new `trait API_Common`
 * attempted to update `class Bitbucket_Server_API`, please let me know if I made 💩
+* refactor release asset and AWS download link code
 
 #### 8.3.1 / 2018-09-13
 * created `class Bootstrap` to setup plugin loading